refactor(simulation): split views.py into 3 files
