+++ gsl_simulation/urls.py
@@ -1,36 +1,44 @@
 from django.urls import path
 
-from . import views
+from gsl_simulation.views import simulation_views
+from gsl_simulation.views.decorators import simulation_must_be_visible_by_user
+from gsl_simulation.views.simulation_projet_views import (
+    patch_montant_simulation_projet,
+    patch_status_simulation_projet,
+    patch_taux_simulation_projet,
+)
 
 urlpatterns = [
     path(
         "simulations/",
-        views.SimulationListView.as_view(),
+        simulation_views.SimulationListView.as_view(),
         name="simulation-list",
     ),
     path(
         "simulation/<slug:slug>/",
-        views.simulation_must_be_visible_by_user(views.SimulationDetailView.as_view()),
+        simulation_must_be_visible_by_user(
+            simulation_views.SimulationDetailView.as_view()
+        ),
         name="simulation-detail",
     ),
     path(
         "modifier-le-taux-d-un-projet-de-simulation/<int:pk>/",
-        views.patch_taux_simulation_projet,
+        patch_taux_simulation_projet,
         name="patch-simulation-projet-taux",
     ),
     path(
         "modifier-le-montant-un-projet-de-simulation/<int:pk>/",
-        views.patch_montant_simulation_projet,
+        patch_montant_simulation_projet,
         name="patch-simulation-projet-montant",
     ),
     path(
         "modifier-le-statut-d-un-projet-de-simulation/<int:pk>/",
-        views.patch_status_simulation_projet,
+        patch_status_simulation_projet,
         name="patch-simulation-projet-status",
     ),
     path(
         "creation-simulation",
-        views.simulation_form,
+        simulation_views.simulation_form,
         name="simulation-form",
     ),
 ]
